fix: eliminate cast_precision_loss in visited_count
Change the visit-count map from HashMap<String, usize> to
HashMap<String, u32>. u32 values all convert to f64 without rounding,
so f64::from(count) replaces the lossy 'count as f64' cast.
RunnerSnapshot::visits updated to match.

+++ src/runtime/runner/mod.rs
@@ -58,7 +58,7 @@ pub struct Runner<S: VariableStorage> {
     pub(super) pending: VecDeque<DialogueEvent>,
     pub(super) option_bodies: OptionBodies,
     pub(super) library: FunctionLibrary,
-    pub(super) visits: Arc<Mutex<HashMap<String, usize>>>,
+    pub(super) visits: Arc<Mutex<HashMap<String, u32>>>,
     pub(super) once_seen: HashSet<String>,
     pub(super) saliency: Box<dyn SaliencyStrategy>,
     pub(super) provider: Box<dyn LineProvider>,
@@ -71,7 +71,7 @@ impl<S: VariableStorage> Runner<S> {
     /// Panics only if the internal `Mutex` is poisoned, which cannot happen in normal use.
     #[must_use]
     pub fn new(program: Program, storage: S) -> Self {
-        let visits: Arc<Mutex<HashMap<String, usize>>> = Arc::new(Mutex::new(HashMap::new()));
+        let visits: Arc<Mutex<HashMap<String, u32>>> = Arc::new(Mutex::new(HashMap::new()));
         let mut library = FunctionLibrary::new();
 
         let v1 = Arc::clone(&visits);
@@ -101,8 +101,7 @@ impl<S: VariableStorage> Runner<S> {
                 }
             };
             let count = *v2.lock().unwrap().get(&title).unwrap_or(&0);
-            #[allow(clippy::cast_precision_loss)]
-            Ok(Value::Number(count as f64))
+            Ok(Value::Number(f64::from(count)))
         });
 
         Self {
